requirements.txt
asgiref==3.2.3
Django==2.0.3
djangorestframework==3.11.0
mysqlclient==1.3.13
pytz==2019.3
sqlparse==0.3.0
